# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FinWise Personal Finance API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Expense {
  id: number;
  amount: number;
  category: string;
  description: string;
  date: string;
  paymentMethod: string;
  createdAt: string;
}

export interface ExpenseInput {
  amount: number;
  category: string;
  description: string;
  date: string;
  paymentMethod: string;
}

export interface ExpenseUpdate {
  amount?: number;
  category?: string;
  description?: string;
  date?: string;
  paymentMethod?: string;
}

export interface Budget {
  id: number;
  category: string;
  amount: number;
  month: string;
  spent: number;
  remaining?: number;
}

export interface BudgetInput {
  category: string;
  amount: number;
  month: string;
}

export interface BudgetUpdate {
  amount?: number;
}

export interface AnalyticsSummary {
  todayTotal: number;
  weekTotal: number;
  monthTotal: number;
  topCategory: string;
  savingsEstimate: number;
  recentExpenses: Expense[];
}

export interface DailySpend {
  date: string;
  total: number;
}

export interface CategorySpend {
  category: string;
  total: number;
  count: number;
  percentage: number;
}

export interface SpendingComparison {
  thisWeek: number;
  lastWeek: number;
  thisMonth: number;
  lastMonth: number;
  weekChange: number;
  monthChange: number;
  avgDailyThisMonth?: number;
  avgDailyLastMonth?: number;
}

export interface HealthScore {
  score: number;
  grade: string;
  budgetAdherence: number;
  savingsRate: number;
  spendingConsistency: number;
  message?: string;
}

export interface Insight {
  id: string;
  type: string;
  title: string;
  message: string;
  severity: string;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  savingsPotential?: number | null;
}

export interface ChatHistoryItem {
  role: string;
  content: string;
}

export interface ChatMessageInput {
  message: string;
  history?: ChatHistoryItem[];
}

export interface ChatResponse {
  reply: string;
}

export type ListExpensesParams = {
/**
 * @nullable
 */
startDate?: string | null;
/**
 * @nullable
 */
endDate?: string | null;
/**
 * @nullable
 */
category?: string | null;
/**
 * @nullable
 */
paymentMethod?: string | null;
/**
 * @nullable
 */
limit?: number | null;
/**
 * @nullable
 */
offset?: number | null;
};

export type ListBudgetsParams = {
/**
 * Month in YYYY-MM format
 * @nullable
 */
month?: string | null;
};

export type GetDailyAnalyticsParams = {
startDate: string;
endDate: string;
};

export type GetCategoryAnalyticsParams = {
/**
 * @nullable
 */
startDate?: string | null;
/**
 * @nullable
 */
endDate?: string | null;
};