# lib/api-zod/src/generated/types/analyticsSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FinWise Personal Finance API
 * OpenAPI spec version: 0.1.0
 */
import type { Expense } from './expense';

export interface AnalyticsSummary {
  todayTotal: number;
  weekTotal: number;
  monthTotal: number;
  topCategory: string;
  savingsEstimate: number;
  recentExpenses: Expense[];
}